Refactor backend to centralize game logic
This commit refactors the backend PHP code to improve its structure, reduce redundancy, and enhance maintainability.

Key changes:

1.  **Created `GameUtils.php`:** A new utility file at `backend/utils/GameUtils.php` has been created to serve as a central location for all core game logic.

2.  **Consolidated Game Logic:**
    -   Moved the `dealCards` function from `match.php` into `GameUtils.php`.
    -   Moved all hand evaluation and scoring functions from `sssScorer.php` and `eightCardScorer.php` into `GameUtils.php`.

3.  **Refactored API Endpoints:**
    -   `match.php` and `submit_hand.php` have been updated to remove their local logic and instead include and use the functions from the new `GameUtils.php`.

4.  **Cleaned Up Old Files:**
    -   Deleted the now-redundant `backend/utils/sssScorer.php` and `backend/utils/eightCardScorer.php` files.
    -   Deleted the unused `backend/database` directory and its contents.
