+++ frontend/src/App.css
@@ -1,47 +1,31 @@
-/* 全局深色渐变统一主题与响应式布局 */
+/* 全局深色渐变统一主题 */
 
 .app {
-  min-height: 100vh;
-  min-height: 100dvh;
-  width: 100vw;
-  display: flex;
-  flex-direction: column;
+  height: 100%; width: 100%;
+  display: flex; flex-direction: column;
   align-items: center;
   background: linear-gradient(135deg, #20243b 0%, #18182f 40%, #121223 100%);
-  color: #f5f6fa;
-  padding: 0;
-  box-sizing: border-box;
-  overflow-x: hidden;
+  color: #f5f6fa; padding: 0; box-sizing: border-box;
 }
 
 .app-main {
-  flex: 1;
-  width: 100%;
-  max-width: 550px;
-  min-height: 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-start;
+  flex: 1; width: 100%; max-width: 550px; min-height: 0;
+  display: flex; flex-direction: column;
+  align-items: center; justify-content: flex-start;
   background: linear-gradient(120deg, #23243c 0%, #171730 80%);
   color: #f5f6fa;
   border-radius: 18px;
   box-shadow: 0 4px 32px rgba(20,20,40,0.25);
-  padding: 20px 0;
-  margin-top: 16px;
-  box-sizing: border-box;
+  padding: 10px 0; margin-top: 10px;
 }
 
 .top-banner {
-  width: 100%;
-  max-width: 550px;
-  margin-bottom: 14px;
-  padding: 18px 12px 18px 12px;
+  width: 100%; max-width: 550px;
+  margin-bottom: 14px; padding: 18px;
   border-radius: 18px;
   background: linear-gradient(90deg, #23243c 0%, #34365c 100%);
   box-shadow: 0 2px 16px rgba(30,30,60,0.20);
   color: #f5f6fa;
-  box-sizing: border-box;
 }
 
 .banner-title {
@@ -52,23 +36,14 @@
 }
 .banner-welcome { font-size: 1rem; color: #c3c8e6; }
 .banner-actions {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  gap: 10px;
-  margin-top: 7px;
+  width: 100%; display: flex; justify-content: center; gap: 10px;
 }
 
 /* 按钮统一风格 */
 .action-button, .banner-btn {
-  flex: 1;
-  min-width: 0;
-  padding: 12px 8px;
-  font-size: 1rem;
-  font-weight: 600;
-  border-radius: 12px;
-  border: none;
-  cursor: pointer;
+  flex: 1; padding: 12px 15px;
+  font-size: 1rem; font-weight: 600;
+  border-radius: 12px; border: none; cursor: pointer;
   background: linear-gradient(90deg, #42436c 0%, #1a1a3e 100%);
   color: #e3e6ff;
   box-shadow: 0 2px 10px rgba(30,30,60,0.10);
@@ -76,49 +51,29 @@
 }
 .action-button:hover, .banner-btn:hover {
   background: linear-gradient(90deg, #89f0ff 0%, #23243c 100%);
-  color: #23243c;
-  box-shadow: 0 6px 20px rgba(44,62,80,0.22);
+  color: #23243c; box-shadow: 0 6px 20px rgba(44,62,80,0.22);
 }
 .action-button:active, .banner-btn:active { transform: scale(0.96); }
 
 /* 弹窗统一深色 */
 .modal-backdrop {
-  position: fixed;
-  z-index: 1100;
-  inset: 0;
+  position: fixed; z-index: 1100; inset: 0;
   background: rgba(26, 27, 38, 0.92);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 0;
+  display: flex; align-items: center; justify-content: center; padding: 15px;
 }
-
 .modal-content {
-  width: 96vw;
-  max-width: 420px;
-  max-height: 92vh;
-  overflow-y: auto;
+  width: 100%; max-width: 420px; max-height: 90vh; overflow-y: auto;
   background: linear-gradient(135deg, #23243c 0%, #171730 100%);
-  border-radius: 20px;
-  box-shadow: 0 8px 38px 0 rgba(16, 20, 40, 0.32);
-  padding: 22px 8vw 22px 8vw;
-  color: #f5f6fa;
-  box-sizing: border-box;
+  border-radius: 20px; box-shadow: 0 8px 38px 0 rgba(16, 20, 40, 0.32);
+  padding: 32px; color: #f5f6fa;
 }
-.modal-content h3 { margin: 0 0 15px 0; font-size: 1.2rem; font-weight: 700; color: #89f0ff; text-align: center; }
+.modal-content h3 { margin: 0 0 15px 0; font-size: 1.3rem; font-weight: 700; color: #89f0ff; text-align: center; }
 .release-notes li { margin-bottom: 8px; color: #c3c8e6; font-size: 0.95rem; }
 .modal-actions { margin-top: 20px; display: flex; gap: 15px; }
 .modal-actions button {
-  flex: 1;
-  padding: 13px 0;
-  border: none;
-  border-radius: 12px;
-  font-size: 1rem;
-  font-weight: bold;
-  cursor: pointer;
-  background: #42436c;
-  color: #e3e6ff;
-  box-shadow: 0 2px 8px rgba(44,62,80,0.20);
+  flex: 1; padding: 13px 0; border: none; border-radius: 12px;
+  font-size: 1rem; font-weight: bold; cursor: pointer;
+  background: #42436c; color: #e3e6ff; box-shadow: 0 2px 8px rgba(44,62,80,0.20);
   transition: background 0.2s;
 }
 .modal-actions button:active { transform: scale(0.97); }
@@ -128,25 +83,30 @@
 .update-btn:hover { background: #6ad6e2; color: #23243c; box-shadow: 0 8px 24px #89f0ff44; }
 .error-message { color: #ff6b81; background: #23243c; border-radius: 8px; padding: 10px; font-weight: bold; text-align: center; margin-top: 10px; }
 
-@media (max-width: 800px) {
-  .app-main, .top-banner {
-    max-width: 100vw;
-    border-radius: 0;
-    padding-left: 0; padding-right: 0;
-    margin-top: 0;
+@media (max-width: 480px) {
+  .app-main {
+    padding: 5px 0;
+    margin-top: 5px;
+    border-radius: 10px;
+  }
+  .top-banner {
+    padding: 12px;
+    border-radius: 10px;
+  }
+  .banner-title {
+    font-size: 1.2rem;
+  }
+  .banner-actions {
+    gap: 8px;
+  }
+  .action-button, .banner-btn {
+    padding: 10px;
+    font-size: 0.9rem;
   }
   .modal-content {
-    max-width: 98vw;
-    padding: 15px 3vw;
+    padding: 20px;
   }
-}
-@media (max-width: 480px) {
-  .app-main, .top-banner {
-    max-width: 100vw;
-    border-radius: 0;
-    margin-top: 0;
-    padding: 0 0 0 0;
+  .modal-content h3 {
+    font-size: 1.1rem;
   }
-  .banner-title { font-size: 1.07rem; }
-  .modal-content { padding: 10px 2vw; }
 }
