+++ frontend/src/App.jsx
@@ -7,8 +7,7 @@ import UserProfile from './components/UserProfile';
 import TransferPoints from './components/TransferPoints';
 import ThirteenGame from './components/ThirteenGame';
 import EightCardGame from './components/EightCardGame';
-import PracticeThirteenGame from './components/PracticeThirteenGame';
-import PracticeEightCardGame from './components/PracticeEightCardGame';
+import ModeSelection from './components/ModeSelection';
 import './App.css';
 import { Browser } from '@capacitor/browser';
 
@@ -46,13 +45,10 @@ function App() {
   const [user, setUser] = useState(null);
   const [gameState, setGameState] = useState({ gameType: null, gameMode: null, roomId: null, error: null });
   const [currentView, setCurrentView] = useState('lobby');
-  // --- 独立匹配状态 ---
+  const [selectedGameType, setSelectedGameType] = useState(null);
   const [matchingStatus, setMatchingStatus] = useState({ thirteen: false, eight: false });
   const [updateInfo, setUpdateInfo] = useState({ show: false, version: '', notes: [], url: '' });
   const [showTransfer, setShowTransfer] = useState(false);
-  // 试玩模式相关
-  const [practiceMode, setPracticeMode] = useState(false);
-  const [practiceConfig, setPracticeConfig] = useState(null);
 
   useEffect(() => {
     const storedUser = localStorage.getItem('user');
@@ -74,13 +70,17 @@ function App() {
     localStorage.removeItem('user');
     setUser(null);
     setGameState({ gameType: null, gameMode: null, roomId: null, error: null });
-    setPracticeMode(false);
-    setPracticeConfig(null);
     setMatchingStatus({ thirteen: false, eight: false });
+    setCurrentView('lobby');
+    setSelectedGameType(null);
   };
 
-  // --- 独立匹配入口 ---
-  const handleSelectGame = async (gameType, gameMode) => {
+  const handleSelectGame = (gameType) => {
+    setSelectedGameType(gameType);
+    setCurrentView('modeSelection');
+  };
+
+  const handleSelectMode = async (gameType, gameMode) => {
     if (matchingStatus[gameType] || !user) return;
     setMatchingStatus(prev => ({ ...prev, [gameType]: true }));
     setGameState({ ...gameState, gameType, gameMode, error: null });
@@ -89,6 +89,7 @@ function App() {
       const data = await response.json();
       if (data.success && data.roomId) {
         setGameState({ gameType: gameType, gameMode: gameMode, roomId: data.roomId, error: null });
+        setCurrentView('game');
       } else {
         setMatchingStatus(prev => ({ ...prev, [gameType]: false }));
         setGameState({ ...gameState, error: data.message || '匹配失败，请重试' });
@@ -99,7 +100,6 @@ function App() {
     }
   };
 
-  // --- 匹配轮询，仅针对当前游戏 ---
   useEffect(() => {
     const currentGame = gameState.gameType;
     if (!currentGame || !matchingStatus[currentGame] || !user) return;
@@ -113,7 +113,7 @@ function App() {
         clearInterval(intervalId);
         return;
       }
-      await handleSelectGame(currentGame, gameState.gameMode);
+      await handleSelectMode(currentGame, gameState.gameMode);
     }, 2000);
     return () => clearInterval(intervalId);
   }, [matchingStatus, user, gameState.roomId, gameState.gameType, gameState.gameMode]);
@@ -122,8 +122,7 @@ function App() {
     setGameState({ gameType: null, gameMode: null, roomId: null, error: null });
     setCurrentView('lobby');
     setMatchingStatus({ thirteen: false, eight: false });
-    setPracticeMode(false);
-    setPracticeConfig(null);
+    setSelectedGameType(null);
   };
 
   const handleUpdate = async () => {
@@ -149,10 +148,8 @@ function App() {
     setCurrentView('lobby');
   };
 
-  const isInGame = !!gameState.roomId;
-
   const renderMainContent = () => {
-    if (isInGame) {
+    if (currentView === 'game' && gameState.roomId) {
       const gameProps = {
         roomId: gameState.roomId,
         gameMode: gameState.gameMode,
@@ -163,14 +160,14 @@ function App() {
       if (gameState.gameType === 'thirteen') return <ThirteenGame {...gameProps} />;
       if (gameState.gameType === 'eight') return <EightCardGame {...gameProps} />;
     }
-    if (practiceMode && practiceConfig && user) {
-      if (practiceConfig.gameType === 'thirteen')
-        return <PracticeThirteenGame aiCount={practiceConfig.aiCount} onBackToLobby={handlePracticeEnd} user={user} />;
-      if (practiceConfig.gameType === 'eight')
-        return <PracticeEightCardGame aiCount={practiceConfig.aiCount} onBackToLobby={handlePracticeEnd} user={user} />;
+
+    if (currentView === 'modeSelection' && selectedGameType) {
+        return <ModeSelection gameType={selectedGameType} onSelectMode={handleSelectMode} onBack={handleBackToLobby} />;
     }
+
     if (gameState.error) return <p className="error-message">{gameState.error}</p>;
     if (showTransfer && user) return <TransferPoints fromId={user.id} onClose={() => setShowTransfer(false)} onSuccess={handleTransferSuccess} />;
+
     switch (currentView) {
       case 'profile': return <UserProfile userId={user.id} user={user} onLogout={handleLogout} onTransferClick={() => setShowTransfer(true)} />;
       case 'lobby':
@@ -179,7 +176,6 @@ function App() {
           <GameLobby
             onSelectGame={handleSelectGame}
             matchingStatus={matchingStatus}
-            onPractice={handlePracticeStart}
           />
         );
     }
