+++ frontend/src/components/EightCardGame.css
@@ -1,5 +1,6 @@
 .table-root {
   width: 100vw;
+  height: 100dvh;
   min-height: 100dvh;
   padding: 0;
   margin: 0;
@@ -8,11 +9,11 @@
   justify-content: stretch;
   align-items: stretch;
   box-sizing: border-box;
-  background: linear-gradient(135deg, #1e2136 0%, #23243c 80%);
 }
 
 .table-panel {
   width: 100vw;
+  height: 100dvh;
   min-height: 100dvh;
   max-width: 100vw;
   padding: 0;
@@ -22,179 +23,29 @@
   gap: 0;
   position: relative;
   box-sizing: border-box;
-  background: none;
+  background: linear-gradient(120deg, #23243c 0%, #171730 80%);
   border-radius: 0;
 }
 
-.table-top-bar {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 14px 10px 8px 10px;
-  background: linear-gradient(90deg, #23243c 70%, #171730 100%);
-  border-bottom: 1.5px solid #23243c55;
-  box-shadow: 0 2px 12px #17173022;
-}
-.table-quit-btn {
-  background: #fc5c7d;
-  color: #fff;
-  border: none;
-  border-radius: 12px;
-  font-size: 1.04rem;
-  padding: 11px 24px;
-  font-weight: bold;
-  cursor: pointer;
-  transition: background .18s, transform .2s;
-}
-.table-quit-btn:hover { background: #ff6b81; transform: scale(1.04); }
-.table-score-box {
-  font-size: 1.13rem;
-  color: #89f0ff;
-  font-weight: 700;
-  letter-spacing: 1px;
-  margin-right: 10px;
-}
-
-/* 横向玩家状态栏 */
-.player-bar-horizontal {
-  width: 100vw;
-  max-width: 100vw;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  align-items: stretch;
-  background: linear-gradient(90deg, #191a34 50%, #23243c 100%);
-  border-bottom: 1px solid #23243c33;
-  box-shadow: 0 2px 10px #23243c22;
-  padding: 0 0 0 0;
-  margin: 0 0 7px 0;
-  min-height: 64px;
-}
-.player-bar-item {
-  flex: 1 1 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  min-width: 0;
-  padding: 6px 0;
-  position: relative;
-  font-size: 0.97rem;
-  font-weight: 600;
-  letter-spacing: 0.5px;
-  color: #c3c8e6;
-  opacity: 0.85;
-  border-right: 1px solid #23243c22;
-}
-.player-bar-item:last-child { border-right: none; }
-.player-bar-item.me { color: #fff; font-weight: bold; background: #34365c33; opacity: 1; }
-.player-bar-item.ready { color: #37e6c5; opacity: 1; }
-.player-dot {
-  width: 8px; height: 8px; border-radius: 50%;
-  background: #ffb142;
-  margin-bottom: 3px;
-  box-shadow: 0 1px 7px #ffb14255;
-}
-.player-bar-item.ready .player-dot { background: #27ae60; box-shadow: 0 0 7px #27ae6055; }
-.player-bar-item.me .player-dot { background: #89f0ff; box-shadow: 0 0 7px #89f0ff55; }
-.player-bar-nick {
-  font-size: 1.02rem;
-  font-weight: 700;
-  color: inherit;
-  margin-bottom: 2px;
-  letter-spacing: 0.8px;
-}
-.player-bar-status {
-  font-size: 0.86rem;
-  color: #abb2d8;
-  font-weight: 400;
-}
-
-/* 理牌区 */
 .table-lanes-area {
   flex-grow: 1;
   min-height: 0;
   overflow-y: auto;
   overscroll-behavior: contain;
-  padding-bottom: 78px;
-  padding-top: 12px;
-  background: none;
-  display: flex;
-  flex-direction: column;
-  gap: 14px;
-  align-items: center;
-}
-.lane-wrapper {
-  width: 98vw;
-  max-width: 420px;
-  margin: 0 auto 0 auto;
-  background: #23243c;
-  border-radius: 15px;
-  box-shadow: 0 3px 14px #18182f17;
-  padding: 13px 8px 7px 8px;
+  padding-bottom: 70px; /* 留出底部空间，防止操作栏被裁切 */
 }
 
-@media (max-width: 520px) {
-  .lane-wrapper { max-width: 99vw; }
-  .table-panel, .player-bar-horizontal { max-width: 100vw; }
-}
-
-/* 底部操作区 */
 .table-actions-bar {
-  width: 100vw;
-  left: 0;
   display: flex;
-  gap: 12px;
-  justify-content: center;
-  background: linear-gradient(180deg,#23243c 70%,#171730 100%);
-  padding: 13px 8px 22px 8px;
-  box-sizing: border-box;
-  z-index: 100;
+  gap: 10px;
+  padding: 9px 8px 18px 8px;
   position: sticky;
   bottom: 0;
-  border-top: 1.2px solid #23243c44;
-}
-.action-btn {
-  flex: 1 1 0;
-  min-width: 0;
-  padding: 15px 0;
-  border-radius: 13px;
-  border: none;
-  font-size: 1.07rem;
-  font-weight: bold;
-  cursor: pointer;
-  transition: all .21s;
-  background: linear-gradient(90deg,#23243c 0%,#37e6c5 100%);
-  color: #23243c;
-  box-shadow: 0 2px 10px #6a82fb18;
-  letter-spacing: 1.1px;
-}
-.action-btn.orange {
-  background: linear-gradient(90deg,#fc5c7d 0%,#6a82fb 100%);
-  color: #fff;
-}
-.action-btn.green {
-  background: linear-gradient(90deg,#89f0ff 0%,#37e6c5 100%);
-  color: #23243c;
-}
-.action-btn:disabled {
-  opacity: 0.62;
-  cursor: not-allowed;
-}
-.error-message {
-  color: #ff6b81;
-  background: #23243c;
-  border-radius: 8px;
-  padding: 10px;
-  font-weight: bold;
-  text-align: center;
-  margin: 16px 0 0 0;
-  font-size: 1.03rem;
+  background: linear-gradient(180deg, #23243c 80%, #171730 100%);
+  z-index: 100;
 }
 
 @media (max-width: 450px), (max-height: 600px) {
   .table-panel { padding: 0; gap: 0; }
-  .table-lanes-area { padding-bottom: 70px; gap: 8px; }
-  .player-bar-horizontal { min-height: 54px; }
-  .lane-wrapper { padding: 8px 2px 4px 2px; }
+  .table-lanes-area { padding-bottom: 70px; gap: 4px; }
 }
