+++ frontend/src/components/GameLobby.jsx
@@ -11,7 +11,7 @@ const GameLobby = ({ onSelectGame, matchingStatus, onPractice }) => {
         const response = await fetch('/api/get_announcement.php');
         const data = await response.json();
         if (data.success && data.text) setAnnouncement(data.text);
-      } catch (error) {}
+      } catch (error) { /* ignore */ }
     };
     fetchAnnouncement();
     const intervalId = setInterval(fetchAnnouncement, 30000);
@@ -32,76 +32,55 @@ const GameLobby = ({ onSelectGame, matchingStatus, onPractice }) => {
   }, []);
 
   return (
-    <div className="lobby-root">
-      <div className="lobby-header">
-        <img src="/vite.svg" alt="logo" className="lobby-logo" />
-        <div className="lobby-header-info">
-          <h1 className="lobby-title">牌局中心</h1>
-          <div className="lobby-online">在线人数 <span>{onlineCount !== null ? onlineCount : '-'}</span></div>
+    <div className="lobby-container">
+      <header className="lobby-header">
+        <h1 className="lobby-title">游戏大厅</h1>
+        <p className="lobby-subtitle">选择你的战场，开启荣耀之旅</p>
+        <div style={{ marginTop: 8, fontSize: '1rem', color: '#ff8a80', fontWeight: 500 }}>
+          当前在线人数：{onlineCount !== null ? onlineCount : '...'}
         </div>
-      </div>
+      </header>
 
       {announcement && (
-        <div className="lobby-announcement">
-          <span role="img" aria-label="megaphone">📢</span> {announcement}
+        <div className="announcement-banner">
+          📢 {announcement}
         </div>
       )}
 
-      <div className="lobby-game-list">
-        {/* 十三张 */}
-        <div className="game-card thirteen">
-          <div className="game-card-bg thirteen"></div>
-          <div className="game-card-main">
-            <div className="game-card-title-area">
-              <div className="game-card-title">经典十三张</div>
-              <div className="game-card-desc">策略与运气的巅峰对决</div>
+      <main className="game-card-grid">
+        {/* 十三张卡片 */}
+        <div className="game-card thirteen-bg">
+          <div className="game-card-overlay">
+            <div className="game-content">
+              <h2 className="game-title">经典十三张</h2>
+              <p className="game-description">策略与运气的巅峰对决</p>
             </div>
-            <div className="game-card-actions">
-              <button className="game-btn primary"
-                onClick={() => onSelectGame('thirteen', 'normal')}
-                disabled={matchingStatus.thirteen}
-              >普通场</button>
-              <button className="game-btn secondary"
-                onClick={() => onSelectGame('thirteen', 'double')}
-                disabled={matchingStatus.thirteen}
-              >翻倍场</button>
-              <button className="game-btn ghost"
-                onClick={() => onPractice('thirteen', 3)}
-                disabled={matchingStatus.thirteen}
-              >试玩</button>
+            <div className="game-actions">
+              <button className="btn btn-primary"
+                onClick={() => onSelectGame('thirteen')}
+              >
+                进入游戏
+              </button>
             </div>
           </div>
         </div>
-        {/* 八张 */}
-        <div className="game-card eight">
-          <div className="game-card-bg eight"></div>
-          <div className="game-card-main">
-            <div className="game-card-title-area">
-              <div className="game-card-title">急速八张</div>
-              <div className="game-card-desc">快节奏的竞技体验</div>
+        {/* 八张卡片 */}
+        <div className="game-card eight-bg">
+          <div className="game-card-overlay">
+            <div className="game-content">
+              <h2 className="game-title">急速八张</h2>
+              <p className="game-description">快节奏的竞技体验</p>
             </div>
-            <div className="game-card-actions">
-              <button className="game-btn primary"
-                onClick={() => onSelectGame('eight', 'normal')}
-                disabled={matchingStatus.eight}
-              >普通场</button>
-              <button className="game-btn secondary"
-                onClick={() => onSelectGame('eight', 'special')}
-                disabled={matchingStatus.eight}
-              >独头场</button>
-              <button className="game-btn ghost"
-                onClick={() => onPractice('eight', 5)}
-                disabled={matchingStatus.eight}
-              >试玩</button>
+            <div className="game-actions">
+              <button className="btn btn-primary"
+                onClick={() => onSelectGame('eight')}
+              >
+                进入游戏
+              </button>
             </div>
           </div>
         </div>
-      </div>
-
-      <footer className="lobby-footer">
-        <span>© 2025 牌局中心</span>
-        <span>wen9521</span>
-      </footer>
+      </main>
     </div>
   );
 };
