+++ frontend/src/components/Lane.css
@@ -1,67 +1,65 @@
 .lane-wrapper {
   width: 100%;
-  margin: 0 auto 20px auto;
-  background: #fff;
-  border-radius: 16px;
-  box-shadow: 0 2px 18px #c9d6ff22;
-  padding: 12px 9px 18px 9px;
   display: flex;
   flex-direction: column;
-  gap: 8px;
-  border: 1.5px solid #e6e9f5;
-}
-
-.lane-header {
-  padding-left: 5px;
-  font-size: 1.08rem;
-  color: #6a82fb;
-  font-weight: 700;
-  margin-bottom: 2px;
-  letter-spacing: 1.1px;
+  gap: 10px; /* 增大区间隔，防止拥挤 */
+  flex: 1;
+  min-height: 0;
 }
 
 .card-placement-box {
   width: 100%;
-  min-height: 85px;
-  background: none;
-  border-radius: 13px;
+  box-sizing: border-box;
+  min-height: 150px; /* 增加高度以容纳更大的卡牌 */
+  background: #171730;
+  border-radius: 14px;
+  box-shadow: 0 1px 4px #23243c22;
   display: flex;
   align-items: flex-end;
   overflow-x: auto;
-  overflow-y: visible !important;
-  padding: 0 8px;
+  overflow-y: visible !important; /* 强力保证弹起不裁切 */
+  padding: 0 12px;
   cursor: pointer;
   position: relative;
 }
+
 .card-wrapper {
   position: relative;
-  transition: transform 0.2s, box-shadow 0.2s;
-  width: 62px;
-  height: 86px;
-  margin-left: -13px;
+  transition: transform 0.2s ease-out, box-shadow 0.2s;
+  width: 80px; /* 增加卡牌宽度 */
+  height: 112px; /* 增加卡牌高度 */
+  margin-left: -53px; /* 调整重叠，保持约1/3的宽度可见 */
 }
+
+/* 堆叠z-index：后面的牌层级高，弹起牌层级最高 */
 .card-wrapper.selected {
   z-index: 99 !important;
-  box-shadow: 0 7px 22px #6a82fb44,0 0 0 2px #6a82fb;
-  transform: translateY(-17px) scale(1.08);
+  box-shadow: 0 7px 22px #00cec9aa,0 0 0 2px #00cec9;
+  transform: translateY(-20px) scale(1.08);
 }
 .card-wrapper.selected .card {
-  border: 2px solid #6a82fb;
-  box-shadow: 0 4px 18px #6a82fb33;
-  filter: brightness(1.12) drop-shadow(0 0 4px #6a82fb);
+  border: 2px solid #00cec9;
+  box-shadow: 0 4px 18px #00cec9b3;
+  filter: brightness(1.15) drop-shadow(0 0 4px #00cec9);
 }
+
 .card-wrapper:first-child { margin-left: 0; }
+
 .card {
   box-sizing: border-box;
-  border-radius: 7px;
+  border-radius: 9px;
   background: #fff;
   width: 100%;
   height: 100%;
 }
 
-@media (max-width: 450px) {
-  .lane-wrapper { padding: 8px 2px 13px 2px; }
-  .card-wrapper { width: 42px; min-width: 32px; max-width: 42px; height: 56px; margin-left: -7px; }
-  .card-placement-box { min-height: 38px; padding: 0 2px; }
-  .lane-header { padding-left: 3px; font-size: 0.93rem; }
+@media (max-width: 480px) {
+  .card-wrapper {
+    width: 64px;
+    height: 90px;
+    margin-left: -42px;
+  }
+  .card-placement-box {
+    min-height: 120px;
+  }
 }
