+++ frontend/src/components/Lane.jsx
@@ -29,26 +29,28 @@ const Lane = ({
     }
   };
 
-  const cardWidth = 60;
-  const defaultOverlap = cardWidth / 2.2;
+  const cardWidth = 80;
+  const defaultOverlap = cardWidth / 3;
   let marginLeft = -defaultOverlap;
 
   if (cards.length > 1 && laneWidth > 0) {
     const totalCardWidth = cards.length * cardWidth;
     const totalOverlappedWidth = (cards.length - 1) * defaultOverlap;
     const totalWidth = totalCardWidth - totalOverlappedWidth;
+
     if (totalWidth > laneWidth) {
       const newOverlap = (totalCardWidth - laneWidth) / (cards.length - 1);
       marginLeft = -newOverlap;
     }
   }
 
+
   return (
-    <div className="lane-wrapper" style={{ overflow: 'visible' }}>
+    <div className="lane-wrapper">
       <div className="lane-header">
         <span className="lane-title">{`${title} (${expectedCount})`}</span>
       </div>
-      <div className="card-placement-box" ref={laneRef} onClick={handleAreaClick} style={{ overflow: 'visible' }}>
+      <div className="card-placement-box" ref={laneRef} onClick={handleAreaClick}>
         {cards.map((card, idx) => {
           const isSelected = selectedCards.some(sel => areCardsEqual(sel, card));
           return (
