+++ frontend/src/components/ModeSelection.css
@@ -1,0 +1,52 @@
+.mode-selection-container {
+    width: 100%;
+    padding: 20px;
+    box-sizing: border-box;
+    text-align: center;
+}
+
+.back-to-lobby-btn {
+    position: absolute;
+    top: 20px;
+    left: 20px;
+    background: none;
+    border: 1px solid #00f6ff;
+    color: #00f6ff;
+    padding: 8px 12px;
+    border-radius: 8px;
+    cursor: pointer;
+    font-size: 0.9rem;
+}
+
+.mode-selection-title {
+    font-size: 1.8rem;
+    color: #e0e0ff;
+    margin-bottom: 30px;
+}
+
+.mode-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+    gap: 20px;
+}
+
+.mode-card {
+    background: rgba(0, 246, 255, 0.1);
+    border: 1px solid #00f6ff;
+    border-radius: 10px;
+    padding: 30px 20px;
+    cursor: pointer;
+    transition: all 0.3s ease;
+}
+
+.mode-card:hover {
+    background: #00f6ff;
+    color: #0d0d2b;
+    transform: translateY(-5px);
+    box-shadow: 0 0 15px #00f6ff;
+}
+
+.mode-card-name {
+    font-size: 1.2rem;
+    font-weight: 600;
+}
