--- frontend/src/components/PracticeThirteenGame.css
@@ -1,220 +1,0 @@
-/* 背景全屏 */
-.pastel-bg {
-    min-height: 100vh;
-    min-height: 100dvh;
-    width: 100vw;
-    background: linear-gradient(135deg, #f5f7fa 0%, #e6eafc 60%, #f9e7fa 100%);
-    display: flex;
-    flex-direction: column;
-    align-items: stretch;
-    box-sizing: border-box;
-    padding: 0;
-    overflow-x: hidden;
-  }
-  
-  /* 主容器100vw/100vh，绝不居中卡片！ */
-  .table-panel {
-    width: 100vw;
-    min-height: 100vh;
-    min-height: 100dvh;
-    max-width: 100vw;
-    background: none;
-    box-shadow: none;
-    display: flex;
-    flex-direction: column;
-    align-items: stretch;
-    padding-bottom: 100px;
-    margin: 0;
-  }
-  
-  /* 顶部栏/返回按钮 */
-  .table-top-bar-v2 {
-    width: 100vw;
-    display: flex;
-    align-items: center;
-    padding: 24px 0 0 0;
-    background: none;
-    justify-content: flex-start;
-    gap: 0;
-  }
-  .table-quit-btn-v2 {
-    background: #ffb6c1;
-    color: #fff;
-    border: none;
-    border-radius: 14px;
-    font-size: 1.1rem;
-    padding: 13px 32px;
-    font-weight: bold;
-    cursor: pointer;
-    transition: background .18s, transform .2s;
-    box-shadow: 0 2px 12px #fcb6c122;
-    margin-left: 4vw;
-  }
-  .table-quit-btn-v2:hover { background: #fc8eac; transform: scale(1.04); }
-  .table-title-v2 {
-    font-size: 1.15rem;
-    font-weight: bold;
-    color: #6a82fb;
-    letter-spacing: 1.2px;
-    margin-left: 18px;
-    flex: 1;
-    text-align: left;
-  }
-  
-  /* 玩家横幅 贴边 */
-  .player-bar-horizontal-v2 {
-    width: 100vw;
-    display: flex;
-    flex-direction: row;
-    background: #f7faff;
-    border-radius: 0 0 0 0;
-    min-height: 52px;
-    box-shadow: 0 2px 12px #e6e9f555;
-    padding: 0;
-    margin: 0;
-    overflow-x: auto;
-  }
-  .player-bar-item-v2 {
-    flex: 1 1 0;
-    min-width: 0;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    padding: 7px 2px 5px 2px;
-    font-size: 0.97rem;
-    font-weight: 600;
-    color: #abb2d8;
-    opacity: 0.88;
-    border-right: 1px solid #e6e9f5;
-  }
-  .player-bar-item-v2:last-child { border-right: none; }
-  .player-bar-item-v2.me { color: #6a82fb; font-weight: bold; background: #eaf0fa; opacity: 1; }
-  .player-bar-item-v2.ready { color: #37e6c5; opacity: 1; }
-  .player-bar-name-v2 { font-size: 1.04rem; font-weight: 700; margin-bottom: 2px; letter-spacing: 0.8px; }
-  .player-bar-status-v2 { font-size: 0.9rem; color: #abb2d8; font-weight: 400; }
-  
-  /* 牌区全屏无边距 */
-  .lane-list-area-v2 {
-    width: 100vw;
-    display: flex;
-    flex-direction: column;
-    gap: 18px;
-    align-items: stretch;
-    justify-content: flex-start;
-    min-height: 0;
-    padding: 18px 0 0 0;
-  }
-  
-  /* 牌墩卡片区 贴边不溢出 */
-  .lane-wrapper {
-    width: 100vw;
-    background: #fff;
-    border-radius: 0 20px 20px 0;
-    box-shadow: 0 3px 24px #c9d6ff22;
-    padding: 18px 2vw 32px 2vw;
-    margin: 0;
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-    border: none;
-    min-height: 90px;
-    box-sizing: border-box;
-  }
-  .lane-header {
-    padding-left: 3px;
-    font-size: 1.04rem;
-    color: #6a82fb;
-    font-weight: 700;
-    margin-bottom: 7px;
-    letter-spacing: 1.1px;
-  }
-  .card-placement-box {
-    width: 100%;
-    min-height: 78px;
-    background: none;
-    border-radius: 13px;
-    display: flex;
-    align-items: flex-end;
-    overflow-x: auto;
-    overflow-y: visible !important;
-    padding: 0 0px;
-    cursor: pointer;
-    position: relative;
-    margin-top: 4px;
-    margin-bottom: 2px;
-  }
-  .card-wrapper {
-    position: relative;
-    transition: transform 0.25s, box-shadow 0.25s;
-    width: 56px;
-    height: 82px;
-    margin-left: -13px;
-    will-change: transform;
-  }
-  .card-wrapper.selected {
-    z-index: 99 !important;
-    box-shadow: 0 10px 32px #a4baff55,0 0 0 2px #6a82fb;
-    transform: translateY(-32px) scale(1.12);
-  }
-  .card-wrapper.selected .card {
-    border: 2px solid #6a82fb;
-    box-shadow: 0 4px 18px #6a82fb33;
-    filter: brightness(1.12) drop-shadow(0 0 4px #6a82fb);
-  }
-  .card-wrapper:first-child { margin-left: 0; }
-  .card {
-    box-sizing: border-box;
-    border-radius: 8px;
-    background: #fff;
-    width: 100%;
-    height: 100%;
-  }
-  
-  /* 底部操作栏全屏贴边吸底 */
-  .footer-action-bar-v2 {
-    position: fixed;
-    left: 0;
-    bottom: 0;
-    width: 100vw;
-    display: flex;
-    gap: 10px;
-    background: linear-gradient(90deg,#f9e7fa 0%,#eaf0fa 100%);
-    box-shadow: 0 -3px 24px #e6e9f555;
-    padding: 13px 2vw 10px 2vw;
-    z-index: 1000;
-    border-top: 1.5px solid #e6e9f5;
-    justify-content: center;
-  }
-  .footer-btn-v2 {
-    flex: 1 1 0;
-    min-width: 0;
-    padding: 14px 0;
-    border-radius: 16px;
-    border: none;
-    font-size: 1.13rem;
-    font-weight: bold;
-    cursor: pointer;
-    transition: all .21s;
-    background: linear-gradient(90deg,#ffb199 0%,#ff8a9a 100%);
-    color: #fff;
-    box-shadow: 0 2px 15px #ffb19922;
-    letter-spacing: 1.1px;
-  }
-  .footer-btn-v2.blue {
-    background: linear-gradient(90deg,#6a82fb 0%,#89f0ff 100%);
-    color: #fff;
-  }
-  .footer-btn-v2.orange {
-    background: linear-gradient(90deg,#ffb199 0%,#ff8a9a 100%);
-    color: #fff;
-  }
-  .footer-btn-v2:active { transform: scale(0.97); }
-  
-  @media (max-width: 600px) {
-    .table-top-bar-v2 { padding-top: 10px; }
-    .lane-wrapper { padding: 10px 1vw 18px 1vw; }
-    .footer-btn-v2 { font-size: 1.01rem; border-radius: 12px; }
-    .footer-action-bar-v2 { padding: 8px 1vw 8vw 1vw; }
-    .player-bar-horizontal-v2 { min-height: 42px; }
-  }
