+++ frontend/src/components/ThirteenGame.jsx
@@ -1,144 +1,227 @@
-import React, { useState, useEffect } from 'react';
-import Lane from './Lane';
-import GameResultModal from './GameResultModal';
-import { getSmartSortedHand } from '../utils/autoSorter';
+// --- START OF FILE frontend/src/components/ThirteenGame.jsx (准备后发牌版) ---
 
-const LANE_LIMITS = { top: 3, middle: 5, bottom: 5 };
-const areCardsEqual = (a, b) => a.rank === b.rank && a.suit === b.suit;
+import React, { useState, useEffect } from 'react';
+import Card from './Card';
+import Lane from './Lane';
+import './ThirteenGame.css';
+import { getSmartSortedHand } from '../utils/autoSorter';
+import { sortCards } from '../utils/pokerEvaluator';
+import GameResultModal from './GameResultModal';
+
+const areCardsEqual = (card1, card2) => {
+  if (!card1 || !card2) return false;
+  return card1.rank === card1.rank && card1.suit === card2.suit;
+};
 
 const ThirteenGame = ({ roomId, gameMode, onBackToLobby, user, onGameEnd }) => {
-  // --- 状态管理 ---
+  const LANE_LIMITS = { top: 3, middle: 5, bottom: 5 };
+
+  // --- 本地状态 ---
   const [topLane, setTopLane] = useState([]);
   const [middleLane, setMiddleLane] = useState([]);
   const [bottomLane, setBottomLane] = useState([]);
   const [selectedCards, setSelectedCards] = useState([]);
-  const [hasDealt, setHasDealt] = useState(false);
+  const [hasDealt, setHasDealt] = useState(false); // 是否已发牌
+  const [isPreparing, setIsPreparing] = useState(false); // 是否已点击准备
+  const [isReady, setIsReady] = useState(false); // 是否已提交理牌
   const [players, setPlayers] = useState([]);
+  const [gameStatus, setGameStatus] = useState('matching'); // 牌桌状态
   const [gameResult, setGameResult] = useState(null);
-  const [isReady, setIsReady] = useState(false);
   const [errorMessage, setErrorMessage] = useState('');
   const [isLoading, setIsLoading] = useState(false);
 
-  // --- 轮询房间状态 ---
+  // --- 轮询游戏状态 ---
   useEffect(() => {
-    let intervalId;
-    const fetchStatus = async () => {
+    if (gameStatus === 'finished') return;
+
+    const intervalId = setInterval(async () => {
       try {
-        const resp = await fetch(`/api/game_status.php?roomId=${roomId}&userId=${user.id}`);
-        const data = await resp.json();
+        const response = await fetch(`/api/game_status.php?roomId=${roomId}&userId=${user.id}`);
+        const data = await response.json();
         if (data.success) {
+          setGameStatus(data.gameStatus);
           setPlayers(data.players);
+
+          // --- 1. 检查自己是否已准备 ---
+          const me = data.players.find(p => p.id === user.id);
+          setIsPreparing(me ? !!me.is_ready : false);
+
+          // --- 2. 检查是否已发牌 ---
           if (data.hand && !hasDealt) {
-            setTopLane(data.hand.top); setMiddleLane(data.hand.middle); setBottomLane(data.hand.bottom);
-            setHasDealt(true); setIsReady(false);
+            setTopLane(data.hand.top);
+            setMiddleLane(data.hand.middle);
+            setBottomLane(data.hand.bottom);
+            setHasDealt(true);
+            setIsReady(false); // 理牌还没提交
           }
+
+          // --- 3. 结果展示 ---
           if (data.gameStatus === 'finished' && data.result) {
             setGameResult(data.result);
             clearInterval(intervalId);
           }
         }
-      } catch {
-        setErrorMessage('网络异常');
+      } catch (error) {
+        setErrorMessage("与服务器断开连接");
         clearInterval(intervalId);
       }
-    };
-    fetchStatus();
-    intervalId = setInterval(fetchStatus, 1200);
+    }, 1000);
     return () => clearInterval(intervalId);
-  }, [roomId, user.id, hasDealt]);
+  }, [roomId, user.id, gameStatus, hasDealt]);
 
-  // --- 玩家横幅平铺 ---
-  const renderPlayerBar = () => (
-    <div className="player-bar-horizontal-v2">
-      {players.map((p, idx) => (
-        <div key={p.id} className={`player-bar-item-v2${p.id === user.id ? ' me' : ''}${p.is_ready ? ' ready' : ''}`}>
-          <div className="player-bar-name-v2">{p.id === user.id ? '你' : (p.phone ? `玩家${p.phone.slice(-4)}` : `玩家${idx + 1}`)}</div>
-          <div className="player-bar-status-v2">{p.is_ready ? '已准备' : '理牌中...'}</div>
-        </div>
-      ))}
-    </div>
-  );
-
-  // --- 卡牌点击/理牌 ---
-  const handleCardClick = (card) => {
-    const selected = selectedCards.some(c => areCardsEqual(c, card));
-    setSelectedCards(selected ? selectedCards.filter(c => !areCardsEqual(c, card)) : [...selectedCards, card]);
-  };
-  const handleLaneClick = (lane) => {
-    if (!selectedCards.length) return;
-    let newTop = topLane.filter(c => !selectedCards.some(sel => areCardsEqual(sel, c)));
-    let newMiddle = middleLane.filter(c => !selectedCards.some(sel => areCardsEqual(sel, c)));
-    let newBottom = bottomLane.filter(c => !selectedCards.some(sel => areCardsEqual(sel, c)));
-    if (lane === 'top') newTop = [...newTop, ...selectedCards];
-    if (lane === 'middle') newMiddle = [...newMiddle, ...selectedCards];
-    if (lane === 'bottom') newBottom = [...newBottom, ...selectedCards];
-    setTopLane(newTop); setMiddleLane(newMiddle); setBottomLane(newBottom); setSelectedCards([]);
-  };
-  const handleAutoSort = () => {
-    const hand = [...topLane, ...middleLane, ...bottomLane];
-    const sorted = getSmartSortedHand(hand);
-    if (sorted) {
-      setTopLane(sorted.top); setMiddleLane(sorted.middle); setBottomLane(sorted.bottom);
-      setSelectedCards([]);
-    }
-  };
-  // --- 提交 ---
-  const handleConfirm = async () => {
-    if (isLoading || isReady) return;
-    if (topLane.length !== LANE_LIMITS.top || middleLane.length !== LANE_LIMITS.middle || bottomLane.length !== LANE_LIMITS.bottom)
-      return setErrorMessage('牌道数量错误！');
+  // --- 1. 点击准备 ---
+  const handlePrepare = async () => {
+    if (isPreparing) return;
     setIsLoading(true);
     setErrorMessage('');
     try {
-      const payload = { userId: user.id, roomId, hand: { top: topLane, middle: middleLane, bottom: bottomLane } };
       const resp = await fetch('/api/player_ready.php', {
-        method: 'POST', headers: { 'Content-Type': 'application/json' }, body: JSON.stringify(payload)
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ userId: user.id, roomId })
       });
       const data = await resp.json();
-      if (data.success) setIsReady(true); else setErrorMessage(data.message || '提交失败');
-    } catch {
-      setErrorMessage('网络异常');
+      if (data.success) {
+        setIsPreparing(true); // 已准备，等发牌
+      } else {
+        setErrorMessage(data.message || '准备失败');
+      }
+    } catch (err) {
+      setErrorMessage('与服务器通信失败');
     } finally {
       setIsLoading(false);
     }
   };
-  const handleCloseResult = () => { setGameResult(null); onBackToLobby(); };
 
+  // --- 2. 理牌提交 ---
+  const handleConfirm = async () => {
+    if (isLoading || isReady) return;
+    if (topLane.length !== LANE_LIMITS.top || middleLane.length !== LANE_LIMITS.middle || bottomLane.length !== LANE_LIMITS.bottom) {
+      setErrorMessage(`牌道数量错误！`);
+      return;
+    }
+    setIsLoading(true);
+    setErrorMessage('');
+    try {
+      const payload = {
+        userId: user.id,
+        roomId: roomId,
+        hand: { top: topLane, middle: middleLane, bottom: bottomLane },
+      };
+      const response = await fetch('/api/player_ready.php', {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify(payload),
+      });
+      const data = await response.json();
+      if (data.success) {
+        setIsReady(true);
+      } else {
+        setErrorMessage(data.message || '提交失败');
+      }
+    } catch (err) {
+      setErrorMessage('与服务器通信失败');
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  // --- 自动理牌 ---
+  const handleAutoSort = () => {
+    if (!hasDealt) return;
+    const sorted = getSmartSortedHand([...topLane, ...middleLane, ...bottomLane]);
+    if (sorted) {
+      setTopLane(sorted.top);
+      setMiddleLane(sorted.middle);
+      setBottomLane(sorted.bottom);
+    }
+  };
+
+  const handleCloseResult = () => {
+    setGameResult(null);
+    onBackToLobby();
+  };
+
+  // --- 未发牌时 UI（准备区） ---
   if (!hasDealt) {
     return (
-      <div className="table-root pastel-bg">
+      <div className="table-root">
         <div className="table-panel">
-          <div className="table-top-bar-v2">
-            <button onClick={onBackToLobby} className="table-quit-btn-v2">退出游戏</button>
-            <div className="table-title-v2">十三张对局</div>
+          <div className="table-top-bar">
+            <button onClick={onBackToLobby} className="table-quit-btn">退出游戏</button>
+            <div className="table-score-box">十三张</div>
           </div>
-          {renderPlayerBar()}
-          <div className="waiting-text">等待所有玩家准备后开始游戏...</div>
+          <div className="players-status-bar">
+            {players.map(p => (
+              <div key={p.id} className={`player-status-item ${p.is_ready ? 'ready' : ''} ${p.id === user.id ? 'you' : ''}`}>
+                <span className="player-name">{p.id === user.id ? '你' : `玩家${p.phone.slice(-4)}`}</span>
+                <span className="status-text">{p.is_ready ? '已准备' : '未准备'}</span>
+              </div>
+            ))}
+          </div>
+          <div style={{ textAlign: 'center', margin: '24px 0', fontSize: '1.25rem', color: '#3b3b8e' }}>
+            等待所有玩家准备后开始游戏...
+          </div>
+          <div style={{ textAlign: 'center' }}>
+            {!isPreparing ? (
+              <button className="action-btn green" onClick={handlePrepare} disabled={isLoading}>
+                {isLoading ? '准备中...' : '点击准备'}
+              </button>
+            ) : (
+              <div style={{ color: '#27ae60', fontWeight: 'bold' }}>你已准备，等待其他玩家...</div>
+            )}
+          </div>
+          {errorMessage && <p className="error-message">{errorMessage}</p>}
         </div>
       </div>
     );
   }
 
+  // --- 已发牌后，显示理牌区 ---
   return (
-    <div className="table-root pastel-bg">
+    <div className="table-root">
       <div className="table-panel">
-        <div className="table-top-bar-v2">
-          <button onClick={onBackToLobby} className="table-quit-btn-v2">退出游戏</button>
-          <div className="table-title-v2">{gameMode === 'double' ? '十三张翻倍场' : '十三张普通场'}</div>
+        <div className="table-top-bar">
+          <button onClick={onBackToLobby} className="table-quit-btn">退出游戏</button>
+          <div className="table-score-box">{gameMode === 'double' ? '十三张翻倍场' : '十三张普通场'}</div>
         </div>
-        {renderPlayerBar()}
-        <div className="lane-list-area-v2">
-          <Lane title="头道" cards={topLane} onCardClick={handleCardClick} onLaneClick={() => handleLaneClick('top')} selectedCards={selectedCards} expectedCount={LANE_LIMITS.top} />
-          <Lane title="中道" cards={middleLane} onCardClick={handleCardClick} onLaneClick={() => handleLaneClick('middle')} selectedCards={selectedCards} expectedCount={LANE_LIMITS.middle} />
-          <Lane title="尾道" cards={bottomLane} onCardClick={handleCardClick} onLaneClick={() => handleLaneClick('bottom')} selectedCards={selectedCards} expectedCount={LANE_LIMITS.bottom} />
+        <div
+          className="players-status-bar"
+          style={{
+            display: 'grid',
+            gridTemplateColumns: 'repeat(auto-fill, minmax(120px, 1fr))',
+            gap: '10px',
+            padding: '10px'
+          }}
+        >
+          {players.map(p => (
+            <div
+              key={p.id}
+              className={`player-status-item ${p.is_ready ? 'ready' : ''} ${p.id === user.id ? 'you' : ''}`}
+              style={{
+                background: 'rgba(0,0,0,0.3)',
+                borderRadius: '8px',
+                padding: '10px',
+                textAlign: 'center'
+              }}
+            >
+              <span className="player-name" style={{display: 'block', fontWeight: 'bold'}}>{p.id === user.id ? `你` : `玩家 ${p.phone.slice(-4)}`}</span>
+              <span className="status-text" style={{fontSize: '0.9em'}}>{p.is_ready ? '已提交' : '理牌中...'}</span>
+            </div>
+          ))}
         </div>
-        <div className="footer-action-bar-v2">
-          <button className="footer-btn-v2 orange" onClick={handleAutoSort} disabled={isReady}>自动理牌</button>
-          <button className="footer-btn-v2 blue" onClick={handleConfirm} disabled={isLoading || isReady}>
+        <div className="table-lanes-area">
+          <Lane title="头道" cards={topLane} onCardClick={null} onLaneClick={null} selectedCards={selectedCards} expectedCount={LANE_LIMITS.top} />
+          <Lane title="中道" cards={middleLane} onCardClick={null} onLaneClick={null} selectedCards={selectedCards} expectedCount={LANE_LIMITS.middle} />
+          <Lane title="尾道" cards={bottomLane} onCardClick={null} onLaneClick={null} selectedCards={selectedCards} expectedCount={LANE_LIMITS.bottom} />
+        </div>
+        {errorMessage && <p className="error-message">{errorMessage}</p>}
+        <div className="table-actions-bar">
+          <button onClick={handleAutoSort} className="action-btn orange" disabled={isReady}>自动理牌</button>
+          <button onClick={handleConfirm} disabled={isLoading || isReady} className="action-btn green">
             {isReady ? '等待其他玩家...' : (isLoading ? '提交中...' : '确认牌型')}
           </button>
         </div>
-        {errorMessage && <div className="table-error">{errorMessage}</div>}
       </div>
       {gameResult && <GameResultModal result={gameResult} onClose={handleCloseResult} />}
     </div>
@@ -146,3 +229,5 @@ const ThirteenGame = ({ roomId, gameMode, onBackToLobby, user, onGameEnd }) => {
 };
 
 export default ThirteenGame;
+
+// --- END OF FILE frontend/src/components/ThirteenGame.jsx (准备后发牌版) ---
