+++ frontend/src/main.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
-import './index.css'
+import './index.css?v=20250820233923'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
